requirements.txt
livekit-agents>=0.8.0
livekit-api>=0.6.0
livekit-plugins-openai>=0.6.0
livekit-plugins-cartesia>=0.6.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-plugins-sarvam>=0.1.0
livekit-plugins-groq>=0.1.0
python-dotenv>=1.0.0
fastapi
uvicorn[standard]
